chore: move esm to its own folder and fix resolving TypeScript typings for ESM build
