+++ esm/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "hashids-esm",
+  "type": "module",
+  "main": "index.js"
+}
