+++ tsconfig.json
@@ -10,5 +10,5 @@
     "declaration": true,
     "strict": true
   },
-  "exclude": ["node_modules", "dist"]
+  "exclude": ["node_modules"]
 }
